% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\docType{class}
\name{Kernel}
\alias{Kernel}
\title{The 'Kernel' class object}
\format{\code{\link{R6Class}} object.}
\usage{
Kernel
}
\value{
an  \code{\link{R6Class}} object which can be used for the rkhs interpolation.
}
\description{
This a abstract class     provide the kernel function and the 1st order derivative of rbf kernel function.
}
\section{Fields}{

\describe{
\item{\code{k_par}}{vector(of length n_hy) containing the hyper-parameter of kernel. n_hy is the length of kernel hyper parameters.}
}}

\section{Methods}{

\describe{
  \item{\code{kern(t1,t2)}}{This method is used to calculate the kernel function given two one dimensional real inputs.}   
  \item{\code{dkd_kpar(t1,t2)}}{This method is used to calculate the gradient of kernel function against the kernel hyper parameters given two one dimensional real inputs.}   
  \item{\code{dkdt(t1,t2)}}{This method is used to calculate the 1st order derivative of kernel function given two one dimensional real inputs.} }
}

\author{
Mu Niu, \email{mu.niu@plymouth.ac.uk}
}
\keyword{data}
