% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.r
\docType{class}
\name{MLP}
\alias{MLP}
\title{The 'MLP' class object}
\format{\code{\link{R6Class}} object.}
\usage{
MLP
}
\value{
an  \code{\link{R6Class}} object which can be used for the rkhs interpolation.
}
\description{
This a R6 class. It inherits from 'kernel' class. It provides the mlp kernel function and the 1st order derivative of mlp kernel function.
}
\author{
Mu Niu, \email{mu.niu@glasgow.ac.uk}
}
\keyword{data}
