% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf.r
\docType{class}
\name{RBF}
\alias{RBF}
\title{The 'RBF' class object}
\format{\code{\link{R6Class}} object.}
\usage{
RBF
}
\value{
an  \code{\link{R6Class}} object which can be used for the rkhs interpolation.
}
\description{
This a R6 class. It inherits from 'kernel' class. It provides the rbf kernel function and the 1st order derivative of rbf kernel function.
}
\author{
Mu Niu, \email{mu.niu@plymouth.ac.uk}
}
\keyword{data}
