% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf.r
\docType{class}
\name{RBF}
\alias{RBF}
\title{The 'RBF' class object}
\format{
\code{\link{R6Class}} object.
}
\value{
an  \code{\link{R6Class}} object which can be used for the rkhs interpolation.
}
\description{
This a R6 class. It inherits from 'kernel' class. It provides the rbf kernel function and the 1st order derivative of rbf kernel function.
}
\author{
Mu Niu, \email{mu.niu@glasgow.ac.uk}
}
\keyword{data}
\section{Super class}{
\code{\link[KGode:Kernel]{KGode::Kernel}} -> \code{RBF}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-greet}{\code{RBF$greet()}}
\item \href{#method-set_k_par}{\code{RBF$set_k_par()}}
\item \href{#method-kern}{\code{RBF$kern()}}
\item \href{#method-dkd_kpar}{\code{RBF$dkd_kpar()}}
\item \href{#method-dkdt}{\code{RBF$dkdt()}}
\item \href{#method-clone}{\code{RBF$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="KGode" data-topic="Kernel" data-id="initialize">}\href{../../KGode/html/Kernel.html#method-initialize}{\code{KGode::Kernel$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-greet"></a>}}
\if{latex}{\out{\hypertarget{method-greet}{}}}
\subsection{Method \code{greet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RBF$greet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_k_par"></a>}}
\if{latex}{\out{\hypertarget{method-set_k_par}{}}}
\subsection{Method \code{set_k_par()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RBF$set_k_par(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kern"></a>}}
\if{latex}{\out{\hypertarget{method-kern}{}}}
\subsection{Method \code{kern()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RBF$kern(t1, t2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dkd_kpar"></a>}}
\if{latex}{\out{\hypertarget{method-dkd_kpar}{}}}
\subsection{Method \code{dkd_kpar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RBF$dkd_kpar(t1, t2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dkdt"></a>}}
\if{latex}{\out{\hypertarget{method-dkdt}{}}}
\subsection{Method \code{dkdt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RBF$dkdt(t1, t2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RBF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
