% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk3g.r
\docType{class}
\name{rkg3}
\alias{rkg3}
\title{The 'rkg3' class object}
\format{\code{\link{R6Class}} object.}
\usage{
rkg3
}
\value{
an  \code{\link{R6Class}} object which can be used for improving ode parameters estimation by using ode as a regularizer.
}
\description{
This class provides advanced gradient matching method by using the ode as a regularizer.
}
\section{Fields}{

\describe{
\item{\code{rk}}{the 'rkhs' class object containing the interpolation information for each state of the ode.}

\item{\code{ode_m}}{the 'ode' class object containing the information about the odes.}
}}

\section{Methods}{

\describe{
  \item{\code{iterate(iter,innerloop,lamb)}}{Iteratively updating ode parameters and interpolation regression coefficients.} 
  \item{\code{witerate(iter,innerloop,dtilda,lamb)}}{Iteratively updating ode parameters and the warped interpolation regression coefficients.}  
  \item{\code{full(par,lam)}}{Updating ode parameters and rkhs interpolation regression coefficients simultaneously. This method is slow but guarantee convergence.} }
}

\author{
Mu Niu, \email{mu.niu@plymouth.ac.uk}
}
\keyword{data}
